Moving components to a separate directory
