+++ src/app/components/add-task/add-task.component.ts
@@ -1,10 +1,10 @@
 import { NgFor } from '@angular/common';
 import { Component, OnInit, TemplateRef, ViewChild } from '@angular/core';
-import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { TaskListComponent } from '../task-list/task-list.component';
-import { TaskService } from '../sevices/task.service';
-import { Task } from '../models/task';
+import { TaskService } from '../../sevices/task.service';
+import { Task } from '../../models/task';
 
 @Component({
   selector: 'app-add-task',
@@ -19,26 +19,26 @@ export class AddTaskComponent implements OnInit {
     "Task"
   ]
   @ViewChild('addformtemplate') popupContent?: TemplateRef<any>;
-  formGroup: FormGroup | null = null;
+  formGroup!: FormGroup;
   constructor(private formBuilder: FormBuilder,
       private dialogRef: MatDialog,
       private taskService: TaskService) { }
 
   ngOnInit(): void {
     this.formGroup = this.formBuilder.group({
-      title: [""],
+      title: ["", [Validators.required,
+        
+      ]],
       description: [""],
       type: [""]
     })
   }
 
   submit() {
-    if (this.formGroup) {
-      let formResult = this.formGroup.value;
-      let task: Task = new Task(formResult.title, formResult.description, formResult.type, "incomplete");
-      this.taskService.addTask(task);
-      this.dialogRef.closeAll();
-    }
+    let formResult = this.formGroup.value;
+    let task: Task = new Task(formResult.title, formResult.description, formResult.type, "incomplete");
+    this.taskService.addTask(task);
+    this.dialogRef.closeAll();
   }
 
   openForm() {
