+++ src/app/components/tasks-card/tasks-card.component.scss
